+++ src/irvine/oeis/a366/A366012.java
@@ -13,6 +13,6 @@ public class A366012 extends LambdaSequence {
 
   /** Construct the sequence. */
   public A366012() {
-    super(0, n -> Integers.SINGLETON.sum(0, n, k -> Binomial.binomial(n + k, k).multiply(Binomial.binomial((long) n * (n + 1), n - k))).divide(n + 1));
+    super(0, n -> Integers.SINGLETON.sum(0, n, k -> Binomial.binomial(n + k, k).multiply(Binomial.binomial(n * (n + 1), n - k))).divide(n + 1));
   }
 }
