+++ src/irvine/oeis/a369/A369439.java
@@ -17,7 +17,7 @@ public class A369439 extends LambdaSequence {
       final long s = 2;
       final long t = 1;
       final long u = 2;
-      return Integers.SINGLETON.sum(0, n / s, k -> Binomial.binomial(t * (n + 1), k).multiply(Binomial.binomial((long) u * (n + 1), n - (long) s * k))).divide(n + 1);
+      return Integers.SINGLETON.sum(0, n / s, k -> Binomial.binomial(t * (n + 1), k).multiply(Binomial.binomial(u * (n + 1), n - s * k))).divide(n + 1);
     });
   }
 }
