+++ src/irvine/oeis/a369/A369440.java
@@ -17,7 +17,7 @@ public class A369440 extends LambdaSequence {
       final long s = 2;
       final long t = 2;
       final long u = 1;
-      return Integers.SINGLETON.sum(0, n / s, k -> Binomial.binomial((long) t * (n + 1), k).multiply(Binomial.binomial(u * (n + 1), n - (long) s * k))).divide(n + 1);
+      return Integers.SINGLETON.sum(0, n / s, k -> Binomial.binomial(t * (n + 1), k).multiply(Binomial.binomial(u * (n + 1), n - s * k))).divide(n + 1);
     });
   }
 }
